pyproject.toml
[project]
name = "document-extractor"
version = "0.1.0"
description = "Extract information from documents (Word, TXT, PDF) and fill forms using AI-powered extraction"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pymupdf>=1.23.0",
    "python-docx>=1.1.0",
    "pytesseract>=0.3.10",
    "langchain>=0.1.0",
    "langchain-google-genai>=1.0.0",
    "playwright>=1.40.0",
    "python-dotenv>=1.0.0",
    "pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []

